Ordinal Higher-Order General Diagnostic Model under the Exploratory Framework (OHOEGDM)

Performs the Gibbs sampling routine for a higher-order EGDM.

ohoegdm(
  y,
  k,
  m = 2,
  order = k,
  sdMH = 0.4,
  burnin = 1000L,
  chain_length = 10000L,
  l0 = c(1, rep(100, sum(choose(k, seq_len(order))))),
  l1 = c(1, rep(1, sum(choose(k, seq_len(order))))),
  m0 = 0,
  bq = 1
)

Arguments

y
Item Matrix

k
Dimension to estimate for Q matrix

m
Number of Item Categories. Default is 2 matching the binary case.

order
Highest interaction order to consider. Default model-specified k.

sdMH
Metropolis-Hastings standard deviation tuning parameter.

burnin
Amount of Draws to Burn

chain_length
Number of Iterations for chain.

l0, l1, m0, bq
Additional tuning parameters.

Value

A egdm_delta object containing three named lists:

estimates

thetas: Average theta coefficients

betas: Average beta coefficients

class: Average class membership

pi: Average attribute class probability.

omega: Average omega

chain

thetas: theta coefficients iterations

betas: beta coefficients iterations

class: class membership iterations

omega: omega iterations

details

n: Number of Subjects

j: Number of Items

k: Number of Traits

order: Highest interaction order to consider. Default model-specified k.

sdMH: Metropolis-Hastings standard deviation tuning parameter.

l0: Spike parameter

l1: Slab parameter

m0, bq: Additional tuning parameters

burnin: Number of Iterations to discard

chain_length: Number of Iterations to keep

Details

The estimates list contains the mean information from the sampling procedure. Meanwhile, the chain list contains full MCMC values. Lastly, the details list provides information regarding the estimation call.

Examples

if (FALSE) {
## Configuration
N = 100  # Number of Subjects
J = 20   # Number of Items
K = 3    # Number of Attributes

## TBA
}